calculator_server.py
from calculator_pb2_grpc import CalculatorServicer

from calculator_pb2 import SumRequest
from calculator_pb2 import SumReply
from calculator_pb2 import MultRequest
from calculator_pb2 import MultReply

from grpc import ServicerContext


class Calculator(CalculatorServicer):

    def Sum(self, request: SumRequest, context: ServicerContext) -> SumReply:
        return SumReply(s=request.a + request.b)
    
    def Mult(self, request: MultRequest, context: ServicerContext) -> MultReply:
        return MultReply(p=request.a * request.b)
