requirements.txt
grpcio==1.63.0
grpcio-tools==1.63.0
iniconfig==2.0.0
packaging==24.0
pluggy==1.5.0
protobuf==5.26.1
pytest==8.2.0
setuptools==69.5.1
